+++ lib/Perl/Tidy/FileWriter.pm
@@ -7,7 +7,7 @@
 package Perl::Tidy::FileWriter;
 use strict;
 use warnings;
-our $VERSION = '20230912.01';
+our $VERSION = '20230912.02';
 
 use constant DEVEL_MODE   => 0;
 use constant EMPTY_STRING => q{};